requirements.txt
pandas>=2.0
shapely>=2.0
requests>=2.30
python-dotenv>=1.0
